chore: update experiment page layout and differentiate multi-single trial [DET-5407]
