+++ webui/react/src/pages/ExperimentDetails.tsx
@@ -1,6 +1,5 @@
-import { Tabs } from 'antd';
 import React, { useCallback, useEffect, useState } from 'react';
-import { useHistory, useParams } from 'react-router';
+import { useParams } from 'react-router';
 
 import CreateExperimentModal, { CreateExperimentType } from 'components/CreateExperimentModal';
 import Message, { MessageType } from 'components/Message';
@@ -8,44 +7,29 @@ import Page from 'components/Page';
 import Spinner from 'components/Spinner';
 import usePolling from 'hooks/usePolling';
 import ExperimentDetailsHeader from 'pages/ExperimentDetails/ExperimentDetailsHeader';
-import ExperimentOverview from 'pages/ExperimentDetails/ExperimentOverview';
 import { paths, routeAll } from 'routes/utils';
 import { getExperimentDetails, getExpValidationHistory, isNotFound } from 'services/api';
 import { createExperiment } from 'services/api';
 import { isAborted } from 'services/utils';
-import { ExperimentBase, ExperimentVisualizationType, RawJson, ValidationHistory } from 'types';
+import { ExperimentBase, ExperimentSearcherName, RawJson, ValidationHistory } from 'types';
 import { clone, isEqual } from 'utils/data';
 import { terminalRunStates, upgradeConfig } from 'utils/types';
 
-const { TabPane } = Tabs;
-
-enum TabType {
-  Configuration = 'configuration',
-  Overview = 'overview',
-  Visualization = 'visualization',
-}
+import ExperimentMultiTrialTabs from './ExperimentDetails/ExperimentMultiTrialTabs';
+import ExperimentSingleTrialTabs from './ExperimentDetails/ExperimentSingleTrialTabs';
 
 interface Params {
   experimentId: string;
-  tab?: TabType;
-  viz?: ExperimentVisualizationType;
 }
 
-const TAB_KEYS = Object.values(TabType);
-const DEFAULT_TAB_KEY = TabType.Overview;
-
-const ExperimentConfiguration = React.lazy(() => {
-  return import('./ExperimentDetails/ExperimentConfiguration');
-});
-const ExperimentVisualization = React.lazy(() => {
-  return import('./ExperimentDetails/ExperimentVisualization');
-});
+export interface ExperimentTabsProps {
+  experiment: ExperimentBase;
+  fetchExperimentDetails: () => void;
+  validationHistory: ValidationHistory[];
+}
 
 const ExperimentDetails: React.FC = () => {
-  const { experimentId, tab, viz } = useParams<Params>();
-  const history = useHistory();
-  const defaultTabKey = tab && TAB_KEYS.includes(tab) ? tab : DEFAULT_TAB_KEY;
-  const [ tabKey, setTabKey ] = useState(defaultTabKey);
+  const { experimentId } = useParams<Params>();
   const [ canceler ] = useState(new AbortController());
   const [ experiment, setExperiment ] = useState<ExperimentBase>();
   const [ valHistory, setValHistory ] = useState<ValidationHistory[]>([]);
@@ -55,7 +39,6 @@ const ExperimentDetails: React.FC = () => {
   const [ isForkModalVisible, setIsForkModalVisible ] = useState(false);
 
   const id = parseInt(experimentId);
-  const basePath = paths.experimentDetails(experimentId);
 
   const fetchExperimentDetails = useCallback(async () => {
     try {
@@ -77,11 +60,6 @@ const ExperimentDetails: React.FC = () => {
   ]);
 
   const { startPolling, stopPolling } = usePolling(fetchExperimentDetails);
-
-  const handleTabChange = useCallback(key => {
-    setTabKey(key);
-    history.replace(key === DEFAULT_TAB_KEY ? basePath : `${basePath}/${key}`);
-  }, [ basePath, history ]);
 
   const showForkModal = useCallback((): void => {
     if (experiment?.configRaw) {
@@ -126,12 +104,6 @@ const ExperimentDetails: React.FC = () => {
   }, [ id, startPolling ]);
 
   useEffect(() => {
-    if (tab && (!TAB_KEYS.includes(tab) || tab === DEFAULT_TAB_KEY)) {
-      history.replace(basePath);
-    }
-  }, [ basePath, history, tab ]);
-
-  useEffect(() => {
     if (experiment && terminalRunStates.has(experiment.state)) {
       stopPolling();
     }
@@ -152,6 +124,10 @@ const ExperimentDetails: React.FC = () => {
     return <Spinner />;
   }
 
+  const isSingleTrial = experiment?.config.searcher.name === ExperimentSearcherName.Single;
+
+  const TabsComponent = isSingleTrial ? ExperimentSingleTrialTabs : ExperimentMultiTrialTabs;
+
   return (
     <Page
       headerComponent={<ExperimentDetailsHeader
@@ -161,27 +137,11 @@ const ExperimentDetails: React.FC = () => {
       />}
       stickyHeader
       title={`Experiment ${experimentId}`}>
-      <Tabs defaultActiveKey={tabKey} onChange={handleTabChange}>
-        <TabPane key="overview" tab="Overview">
-          <ExperimentOverview
-            experiment={experiment}
-            validationHistory={valHistory}
-            onTagsChange={fetchExperimentDetails} />
-        </TabPane>
-        <TabPane key="visualization" tab="Visualization">
-          <React.Suspense fallback={<Spinner />}>
-            <ExperimentVisualization
-              basePath={`${basePath}/${TabType.Visualization}`}
-              experiment={experiment}
-              type={viz} />
-          </React.Suspense>
-        </TabPane>
-        <TabPane key="configuration" tab="Configuration">
-          <React.Suspense fallback={<Spinner />}>
-            <ExperimentConfiguration experiment={experiment} />
-          </React.Suspense>
-        </TabPane>
-      </Tabs>
+      <TabsComponent
+        experiment={experiment}
+        fetchExperimentDetails={fetchExperimentDetails}
+        validationHistory={valHistory}
+      />
       <CreateExperimentModal
         config={forkModalConfig}
         error={forkModalError}
